api/requirements.txt
timm
transformers
fairscale
pycocoevalcap
torch
torchvision
Pillow
scipy
git+https://github.com/openai/CLIP.git
streamlit
fastapi
python-multipart